+++ MooveTeq-Booking/Data/DistanceChosenEventArgs.cs
@@ -6,13 +6,13 @@ using System.Threading.Tasks;
 
 namespace MooveTeqBooking.Data {
     public class DistanceChosenEventArgs : EventArgs {
-        public DistanceChosenEventArgs(bool cancelled, float chosenDistance = 0) {
+        public DistanceChosenEventArgs(bool cancelled, double chosenDistance = 0) {
             Cancelled = cancelled;
             ChosenDistance = chosenDistance;
         }
 
         public bool Cancelled { get; internal set; }
 
-        public float ChosenDistance { get; internal set; }
+        public double ChosenDistance { get; internal set; }
     }
 }
